[2217] Revert temporarily workaround introduced for converting to EntryGuidMap-Storage
Any ideas how to add a safeguard to at least warn about uint64 values used in the uint32 functions are very welcome!

Revert "[2148] Temporarily workaround to restore instances doors or chests"
This reverts commit 4388079dfe2cd3fbed9809e7f4d1db5cc3d9e31d

Revert "[2149] Forgotten part of last commit..."
This reverts commit 3abfecdc88640b83fa90042167ec4ff3504a134f.

+++ include/sc_instance.h
@@ -35,13 +35,11 @@ class MANGOS_DLL_DECL ScriptedInstance : public InstanceData
 
         // Change active state of doors or buttons
         void DoUseDoorOrButton(ObjectGuid guid, uint32 uiWithRestoreTime = 0, bool bUseAlternativeState = false);
-        // FIXME revert back to uint32 uiEntry when no more GO-Guids are stored as uint64
-        void DoUseDoorOrButton(uint64 uiEntry, uint32 uiWithRestoreTime = 0, bool bUseAlternativeState = false);
+        void DoUseDoorOrButton(uint32 uiEntry, uint32 uiWithRestoreTime = 0, bool bUseAlternativeState = false);
 
         // Respawns a GO having negative spawntimesecs in gameobject-table
         void DoRespawnGameObject(ObjectGuid guid, uint32 uiTimeToDespawn = MINUTE);
-        // FIXME revert back to uint32 uiEntry when no more GO-Guids are stored as uint64
-        void DoRespawnGameObject(uint64 uiEntry, uint32 uiTimeToDespawn = MINUTE);
+        void DoRespawnGameObject(uint32 uiEntry, uint32 uiTimeToDespawn = MINUTE);
 
         // Sends world state update to all players in instance
         void DoUpdateWorldState(uint32 uiStateId, uint32 uiStateData);
